chore: rename data-provider dir for MCP
