+++ client/src/data-provider/MCP/queries.ts
