--- client/src/data-provider/McpServers/index.ts
@@ -1,1 +1,0 @@
-export * from './queries';
