--- client/src/data-provider/McpServers/queries.ts
@@ -1,23 +1,0 @@
-import { useQuery, UseQueryOptions, QueryObserverResult } from '@tanstack/react-query';
-import { QueryKeys, dataService } from 'librechat-data-provider';
-import type * as t from 'librechat-data-provider';
-
-/**
- * Hook for fetching all accessible MCP servers with permission metadata
- */
-export const useMCPServersQuery = <TData = t.MCPServersListResponse>(
-  config?: UseQueryOptions<t.MCPServersListResponse, unknown, TData>,
-): QueryObserverResult<TData> => {
-  return useQuery<t.MCPServersListResponse, unknown, TData>(
-    [QueryKeys.mcpServers],
-    () => dataService.getMCPServers(),
-    {
-      staleTime: 1000 * 60 * 5, // 5 minutes - data stays fresh longer
-      refetchOnWindowFocus: false,
-      refetchOnReconnect: false,
-      refetchOnMount: false,
-      retry: false,
-      ...config,
-    },
-  );
-};
