+++ client/src/data-provider/index.ts
@@ -14,4 +14,4 @@ export * from './queries';
 export * from './mcp';
 export * from './roles';
 export * from './tags';
-export * from './MCPServers';
+export * from './MCP';
